requirements.txt
absl-py==1.1.0
astunparse==1.6.3
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
einops==0.2.0
filelock==3.7.1
Flask==2.1.3
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
gdown==4.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
imageio==2.19.3
imgaug==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
libclang==14.0.1
lmdb==1.0.0
Markdown==3.4
MarkupSafe==2.1.1
matplotlib==3.5.2
networkx==2.8.4
numpy==1.23.1
oauthlib==3.2.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
prefetch-generator==1.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.3
scipy==1.8.1
Shapely==1.8.2
six==1.16.0
soupsieve==2.3.2.post1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-serving-api==2.9.1
termcolor==1.1.0
tifffile==2022.5.4
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.10
vietocr==0.3.8
Werkzeug==2.1.2
wrapt==1.14.1
